% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{autoplot.mixpoissonreg}
\alias{autoplot.mixpoissonreg}
\title{autoplot.mixpoissonreg}
\usage{
\method{autoplot}{mixpoissonreg}(
  object,
  which = c(1, 2, 5, 6),
  title = list("Residuals vs Obs. number", "Normal Q-Q", "Cook's distance",
    "Generalized Cook's distance", "Cook's dist vs Generalized Cook's dist",
    "Response vs Fitted means"),
  label.repel = TRUE,
  nrow = NULL,
  ncol = NULL,
  qqline = TRUE,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  include.modeltype = TRUE,
  include.residualtype = FALSE,
  sub.caption = NULL,
  env_alpha = 0.5,
  env_fill = "grey70",
  gpar_sub.caption = list(fontface = "bold"),
  colour = "#444444",
  size = NULL,
  linetype = NULL,
  alpha = NULL,
  fill = NULL,
  shape = NULL,
  label = TRUE,
  label.label = NULL,
  label.colour = "#000000",
  label.alpha = NULL,
  label.size = NULL,
  label.angle = NULL,
  label.family = NULL,
  label.fontface = NULL,
  label.lineheight = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  label.n = 3,
  ad.colour = "#888888",
  ad.linetype = "dashed",
  ad.size = 0.2,
  ...
)
}
\arguments{
\item{env_alpha}{alpha of the bands of the envelope (when the fitted model has envelopes)}

\item{env_fill}{the colour of the filling in the envelopes.}

\item{gpar_sub.caption}{list of gpar parameters to be used as common title in the case of multiple plots. The title will be given in sub.caption argument. See
the help of \code{gpar} function from the \pkg{grid} package for all the available options.}

\item{label.label}{vector of labels.}
}
\value{

}
\description{
autoplot.mixpoissonreg
}
\details{
Based on \code{autoplot.glm} from the excellent \code{ggfortify} package, \\href{https://github.com/sinhrks/ggfortify}.
}
