% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{local_influence_benchmarks.mixpoissonreg}
\alias{local_influence_benchmarks.mixpoissonreg}
\title{local_influence_benchmarks.mixpoissonreg}
\usage{
\method{local_influence_benchmarks}{mixpoissonreg}(
  x,
  perturbation = c("case_weights", "hidden_variable", "mean_explanatory",
    "precision_explanatory", "simultaneous_explanatory"),
  curvature = c("conformal", "normal"),
  direction = c("canonical", "max.eigen"),
  parameters = c("all", "mean", "precision"),
  mean.covariates = NULL,
  precision.covariates = NULL
)
}
\value{

}
\description{
local_influence_benchmarks.mixpoissonreg
}
