% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{local_influence_benchmarks}
\alias{local_influence_benchmarks}
\title{local_influence_benchmarks}
\usage{
local_influence_benchmarks(model, ...)
}
\value{

}
\description{
local_influence_benchmarks
}
