% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_diagnostic_functions.R
\name{logLik.mixpoissonreg}
\alias{logLik.mixpoissonreg}
\title{logLik Method for \code{mixpoissonreg} Objects}
\usage{
\method{logLik}{mixpoissonreg}(object, ...)
}
\arguments{
\item{object}{an object of class "mixpoissonreg" containing results from the fitted model.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns an object of class \code{LogLik} containing the log-likelihood
of the fitted \emph{mixpoissonreg} object.
}
\description{
Function to compute the log-likelihood at the estimated parameters for mixed Poisson regression models.
}
\examples{
daysabs_prog <- mixpoissonreg(daysabs ~ prog, data = Attendance)
logLik(daysabs_prog)

}
\seealso{
\code{\link{vcov.mixpoissonreg}}
}
