% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_auxiliary_functions.R
\name{startvalues_mpreg}
\alias{startvalues_mpreg}
\title{Initial guesses for mixed Poisson regression models}
\usage{
startvalues_mpreg(y, x, w, link.mean, link.precision, model)
}
\arguments{
\item{y}{response vector with z_i >= 0 and integers.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{w}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "log" and "sqrt".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log" or "inverse.sqrt".}

\item{model}{the model, "NB" or "PIG". Currently not used.}
}
\value{
a list containing
\itemize{
\item beta - the initial guesses for the mean-related parameters;
\item alpha - the initial guesses for the precision-relared parameters.
}
@noRd
}
\description{
Function providing initial values for the optimization algorithms for mixed Poisson regression models.
}
