% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{tidy.mixpoissonreg}
\alias{tidy.mixpoissonreg}
\title{tidy.mixpoissonreg}
\usage{
\method{tidy}{mixpoissonreg}(x, conf.int = FALSE, conf.level = 0.95)
}
\value{

}
\description{
tidy.mixpoissonreg
}
