% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_info_datasets.R
\docType{data}
\name{Attendance}
\alias{Attendance}
\title{Attendance Records data set}
\format{
Data frame containing 314 observations on 4 variables.
\describe{
\item{daysabs}{number of days absent.}
\item{gender}{gender of the student.}
\item{prog}{three-level factor indicating the type of instructional program in which the student is enrolled.}
\item{math}{standardized math score.}
}
}
\source{
Data can be obtained from \href{https://github.com/tjfisher19/introStatModeling}{Introduction to Statistical Modeling Github Repository}. See also \emph{Barreto-Souza and Simas (2020)} for further details.
}
\usage{
data("Attendance")
}
\description{
School administrators study the attendance behavior of high school juniors at two schools.
}
\details{
School administrators study the attendance behavior of high school juniors at two schools. Predictors of the number of days of absence include the type of program in which
the student is enrolled and a standardized test in math. Attendance data on 314 high school juniors from two urban high schools.
The response variable of interest is days absent, \code{daysabs}. The variable \code{math} is the standardized math score for each student.
The variable \code{prog} is a three-level factor indicating the type of instructional program in which the student is enrolled.
}
\examples{
\donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math + prog | gender + 
math + prog, data = Attendance, model = "PIG")
summary(daysabs_fit)

daysabs_fit_red <- mixpoissonreg(daysabs ~ gender + math + prog | prog, 
data = Attendance, model = "PIG")
summary(daysabs_fit_red)

# Plot of the fit with all precision covariates
plot(daysabs_fit)
local_influence_plot(daysabs_fit)

# Plot, using ggplot2, of the reduced fit
autoplot(daysabs_fit_red)
local_influence_autoplot(daysabs_fit_red)
}
}
\references{
Hughes, M. and Fisher, T. (2020)  \href{http://www.users.miamioh.edu/fishert4/sta363/}{Introduction to Statistical Modeling.}
}
\keyword{datasets}
