% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{augment.mixpoissonreg}
\alias{augment.mixpoissonreg}
\title{augment.mixpoissonreg}
\usage{
\method{augment}{mixpoissonreg}(
  x,
  data = stats::model.frame(x),
  newdata = NULL,
  type.predict = c("response", "link", "precision", "variance"),
  type.residuals = c("pearson", "score"),
  se_fit = FALSE,
  conf_int = TRUE,
  pred_int = FALSE,
  ...
)
}
\value{

}
\description{
augment.mixpoissonreg
}
