% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_diagnostic_functions.R
\name{logLik.mixpoissonreg}
\alias{logLik.mixpoissonreg}
\title{logLik Method for \code{mixpoissonreg} Objects}
\usage{
\method{logLik}{mixpoissonreg}(object, ...)
}
\arguments{
\item{object}{an object of class "mixpoissonreg" containing results from the fitted model.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Function to compute the log-likelihood at the estimated parameters for mixed Poisson regression models.
}
\examples{
\donttest{
data("Attendance", package = "mixpoissonreg")

daysabs_fit <- mixpoissonreg(daysabs ~ gender + math + 
prog | gender + math + prog, data = Attendance)
logLik(daysabs_fit)
}
}
\seealso{
\code{\link{vcov.mixpoissonreg}}
}
