% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidy_mixpoissonreg.R
\name{tidy_local_influence.mixpoissonreg}
\alias{tidy_local_influence.mixpoissonreg}
\title{tidy_local_influence.mixpoissonreg}
\usage{
\method{tidy_local_influence}{mixpoissonreg}(
  x,
  perturbation = c("case_weights", "hidden_variable", "mean_explanatory",
    "precision_explanatory", "simultaneous_explanatory"),
  curvature = c("conformal", "normal"),
  direction = c("canonical", "max.eigen"),
  parameters = c("all", "mean", "precision"),
  mean.covariates = NULL,
  precision.covariates = NULL
)
}
\value{

}
\description{
tidy_local_influence.mixpoissonreg
}
